requirements.txt
google==3.0.0
google-api-python-client==2.86.0
python-dotenv==1.0.0
boto3==1.26.123